Refactored gin HTTP handlers into handlers.go.
