+++ src/logserver.go
@@ -22,20 +22,9 @@ package main
 import (
 	"logserver/logger"
 
-	"bytes"
-	b64 "encoding/base64"
-	"errors"
-
 	"fmt"
-	"io"
-	"io/ioutil"
 	"log"
-	"mime/multipart"
 	"net/http"
-	"net/url"
-	"os"
-	"path/filepath"
-	"strconv"
 	"time"
 
 	"github.com/gin-gonic/gin"
@@ -50,102 +39,6 @@ var gUseCassandra = false
 func init() {
 	// Todo: Initialize global server logging here.
 	log.Println("[XCONFLOGSERVER] Initializing Xconf Log Server")
-}
-
-// Encode the file as base64 string.
-func encode(buf *bytes.Buffer) string {
-	sEnc := b64.StdEncoding.EncodeToString(buf.Bytes())
-	return sEnc
-}
-
-// Upload the file.
-func upload(context *gin.Context, file *multipart.FileHeader) error {
-	// Open the file for
-	src, err := file.Open()
-	if err != nil {
-		return err
-	}
-	defer src.Close()
-
-	// Read file into buffer.
-	buf := bytes.NewBuffer(nil)
-	if _, err = io.Copy(buf, src); err != nil {
-		return err
-	}
-
-	// Encode file into base64.
-	useEncode := viper.GetBool("logserver.encode")
-	var content []byte
-	if useEncode {
-		sEnc := encode(buf)
-		if sEnc == "" {
-			return errors.New("unable to encode file")
-		}
-		content = []byte(sEnc)
-	} else {
-		content = buf.Bytes()
-	}
-
-	// Send encoded buffer to output URL (specified in config file)
-	filename := filepath.Base(file.Filename)
-	if err = saveLog(content, filename); err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// LogDownloadEntry provides information for a log asset,
-// one that is being downloaded from the log server.
-type LogDownloadEntry struct {
-	Name string      // The name of the log.
-	Path string      // The URL for the log.
-	File os.FileInfo // A reference to the downloaded file.
-}
-
-// Retrieve the logs from the server.
-func getLogs(src string) ([]LogDownloadEntry, error) {
-	var logs []LogDownloadEntry
-	//log.Println("[LOGSERVER-Info] Reading logs from " + src + ".")
-	logger.XconfLogDebug("Retrieving logs from "+src+".", true)
-
-	// The expectation is that the source is specified as a formal URL.
-	u, err := url.Parse(src)
-	if err != nil {
-		//log.Println("[LOGSERVER-Error] Unable to parse source URL " + src + ".")
-		logger.XconfLogError("Unable to parse source URL "+src+".", true)
-		return nil, err
-	}
-
-	if u.Scheme == "file" {
-		// Logs are stored locally on the server.
-		files, err := ioutil.ReadDir(u.Path)
-		if err != nil {
-			//log.Println("[LOGSERVER-Error] Unable to read directory " + u.Path + ".")
-			logger.XconfLogError("Unable to read directory "+u.Path+".", true)
-			return nil, err
-		}
-
-		logs = []LogDownloadEntry{}
-
-		for _, f := range files {
-			//log.Println("[LOGSERVER-Info] Found log file " + f.Name() + ".")
-			logger.XconfLogDebug("Found log file "+f.Name()+".", true)
-			entry := LogDownloadEntry{Name: f.Name(), Path: u.Path, File: f}
-			logs = append(logs, entry)
-		}
-
-	} else if u.Scheme == "http" {
-		//log.Println("[LOGSERVER-Info] Retrieving logs from " + u.Path + ".")
-		logger.XconfLogInfo("Retrieving logs from "+u.Path+".", true)
-		// Todo: retrieve logs from HTTP server.
-	} else {
-		//log.Println("[LOGSERVER-Error] Unsupported source URL " + src + ".")
-		logger.XconfLogError("Unsupported source URL "+src+".", true)
-		return nil, errors.New("unable to upload file")
-	}
-
-	return logs, nil
 }
 
 func main() {
@@ -211,439 +104,16 @@ func main() {
 	})
 
 	// Handler for POST REST API, http://<ip_address>:<port>/api/v1/upload.
-	router.POST("/api/v1/logs/upload", func(c *gin.Context) {
-		// Retrieve form parameters.
-		contact := c.PostForm("contact")
-		description := c.PostForm("description")
-
-		// Retrieve the file from the field "filename".
-		file, err := c.FormFile("filename")
-		if err != nil {
-			c.AbortWithStatusJSON(http.StatusBadRequest, gin.H{
-				"message": "GET Form error",
-				"reason":  err.Error(),
-			})
-			return
-		}
-
-		// Upload the file.
-		err = upload(c, file)
-		if err != nil {
-			c.AbortWithStatusJSON(http.StatusBadRequest, gin.H{
-				"message": "Upload file error",
-				"reason":  err.Error(),
-			})
-			return
-		}
-
-		var entry LogEntry
-		if gUseCassandra {
-			// Update the Cassandra DB if we are using it.
-			entry.FileName = file.Filename
-			entry.Location = viper.GetString("logserver.destination")
-			entry.Size = file.Size
-			entry.Contact = contact
-			entry.Description = description
-
-			var owner string
-			var createDate *time.Time
-			owner, createDate, err = parseFilename(entry.FileName)
-			if err != nil {
-				c.AbortWithStatusJSON(http.StatusInternalServerError, gin.H{
-					"message": "Meta-data registration error",
-					"reason":  err.Error(),
-				})
-				return
-			} else {
-				entry.Owner = owner
-				entry.CreateDate = *createDate
-			}
-
-			// Register the meta-data with the database.
-			err = registerLog(gSession, &entry)
-			if err != nil {
-				//log.Println("[LOGSERVER-Error] Unable to register meta-data for log,", entry.FileName)
-				logger.XconfLogError("Unable to register meta-data for log,"+entry.FileName, true)
-				c.AbortWithStatusJSON(http.StatusInternalServerError, gin.H{
-					"message": "Meta-data registration error",
-					"reason":  err.Error(),
-				})
-				return
-			}
-		}
-
-		var interfaceSlice []interface{} = make([]interface{}, 1)
-		interfaceSlice[0] = entry
-
-		type UploadResponse struct {
-			Message string        `json:"message"`
-			Reason  string        `json:"reason"`
-			Reply   []interface{} `json:"info"`
-		}
-		var response UploadResponse
-		response.Message = "Log upload success"
-		response.Reason = ""
-		response.Reply = interfaceSlice
-
-		c.JSON(http.StatusOK, response)
-	})
+	router.POST("/api/v1/logs/upload", UploadLog)
 
 	// Handler for GET REST API, http://<ip_address>:<port>/api/v1/download?name=<file_name>.
-	router.GET("/api/v1/logs/download", func(c *gin.Context) {
-		//log.Println("[LOGSERVER-Info] In download handler.")
-		logger.XconfLogDebug("In download handler.", true)
-
-		fileName := c.Query("name")
-		if fileName == "" {
-			//log.Println("[LOGSERVER-Error] Query key name was not provided.")
-			logger.XconfLogError("Query key name was not provided.", true)
-			c.AbortWithStatusJSON(http.StatusBadRequest, gin.H{
-				"message": "Query error",
-				"reason":  "name key required",
-			})
-			return
-		}
-
-		src := viper.GetString("logserver.destination")
-		// The expectation is that the source is specified as a formal URL.
-		u, err := url.Parse(src)
-		if err != nil {
-			//log.Println("[LOGSERVER-Error] Unable to parse source URL " + src + ".")
-			logger.XconfLogError("Unable to parse source URL "+src+".", true)
-			c.AbortWithStatusJSON(http.StatusInternalServerError, gin.H{
-				"message": "URL parse error",
-				"reason":  err.Error(),
-			})
-			return
-		}
-
-		if u.Scheme == "file" {
-			//log.Println("[LOGSERVER-Info] Attempting to download log from " + u.Path + ".")
-			logger.XconfLogDebug("Downloading log from "+u.Path+".", true)
-
-			// Validate that the log exists.
-			//path := u.Path + "/" + fileName
-			path := filepath.Join(u.Path, fileName)
-
-			var info os.FileInfo
-			if info, err = os.Stat(path); os.IsNotExist(err) {
-				c.AbortWithStatusJSON(http.StatusNotFound, gin.H{
-					"message": "Download log error",
-					"reason":  err.Error(),
-				})
-				return
-			}
-			len := info.Size()
-			sLen := strconv.FormatInt(len, 10)
-
-			c.Header("Content-Description", "File Transfer")
-			c.Header("Content-Transfer-Encoding", "binary")
-			c.Header("Content-Disposition", "attachment; filename="+fileName)
-			c.Header("Content-Type", "application/octet-stream")
-			c.Header("Content-Length", sLen)
-
-			c.File(path)
-
-			c.JSON(http.StatusOK, gin.H{
-				"message": "Download success",
-				"reason":  "",
-			})
-		} else if u.Scheme == "http" {
-			//log.Println("[LOGSERVER-Info] Downloading log from " + u.Path + ".")
-			logger.XconfLogInfo("Downloading log from "+u.Path+".", true)
-			// Todo: retrieve logs from an HTTP server.
-
-			c.AbortWithStatusJSON(http.StatusUnsupportedMediaType, gin.H{
-				"message": "URL parse error",
-				"reason":  "HTTP protocol not supported",
-			})
-		} else {
-			//log.Println("[LOGSERVER-Error] Unsupported source URL " + src + ".")
-			logger.XconfLogError("Unsupported source URL "+src+".", true)
-
-			c.AbortWithStatusJSON(http.StatusUnsupportedMediaType, gin.H{
-				"message": "URL parse error",
-				"reason":  "protocol not supported",
-			})
-		}
-	})
+	router.GET("/api/v1/logs/download", DownloadLog)
 
 	// Handler for DELETE REST API, http://<ip_address>:<port>/api/v1/logs/<name>.
-	router.DELETE("/api/v1/logs/:name", func(c *gin.Context) {
-		//log.Println("[LOGSERVER-Info] In delete handler.")
-		logger.XconfLogDebug("In delete handler.", true)
-
-		fileName := c.Params.ByName("name")
-		if fileName == "" {
-			//log.Println("[LOGSERVER-Error] Log key name was not provided.")
-			logger.XconfLogError("Log key name was not provided.", true)
-
-			c.AbortWithStatusJSON(http.StatusBadRequest, gin.H{
-				"message": "Delete error",
-				"reason":  "name key required",
-			})
-			return
-		}
-
-		src := viper.GetString("logserver.destination")
-		// The expectation is that the source is specified as a formal URL.
-		u, err := url.Parse(src)
-		if err != nil {
-			log.Println("[LOGSERVER-Error] Unable to parse source URL " + src + ".")
-			logger.XconfLogError("Unable to parse source URL "+src+".", true)
-
-			c.AbortWithStatusJSON(http.StatusInternalServerError, gin.H{
-				"message": "URL parse error",
-				"reason":  err.Error(),
-			})
-			return
-		}
-
-		if u.Scheme == "file" {
-			log.Println("[LOGSERVER-Info] Attempting to delete log from " + u.Path + ".")
-			logger.XconfLogDebug("Deleting log from "+u.Path+".", true)
-
-			// Validate that the log exists.
-			//path := u.Path + "/" + fileName
-			path := filepath.Join(u.Path, fileName)
-
-			if _, err = os.Stat(path); os.IsNotExist(err) {
-				//log.Println("[LOGSERVER-Error] Log does not exist.")
-				logger.XconfLogError("Log does not exist.", true)
-
-				c.AbortWithStatusJSON(http.StatusNotFound, gin.H{
-					"message": "Delete error",
-					"reason":  err.Error(),
-				})
-				return
-			}
-
-			if gUseCassandra {
-				// Remove Cassandra data base entry.
-
-				// Retrieve info for named file.
-				var filter LogFilter
-				filter.FileName = fileName
-				filter.SizeLower = -1
-				filter.SizeUpper = -1
-
-				var info []LogEntry
-				info, err = retrieveLogInfo(gSession, &filter)
-				if err != nil {
-					c.AbortWithStatusJSON(http.StatusInternalServerError, gin.H{
-						"message": "Info retrieval error",
-						"reason":  err.Error(),
-					})
-					return
-				}
-
-				// Unregister the database entry.
-				err = unregisterLog(gSession, &info[0])
-				if err != nil {
-					//log.Println("[LOGSERVER-Error] Unable to unregister log.")
-					logger.XconfLogError("Unable to unregister log.", true)
-					c.AbortWithStatusJSON(http.StatusNotFound, gin.H{
-						"message": "Delete error",
-						"reason":  err.Error(),
-					})
-					return
-				}
-			}
-
-			// Delete the log.
-			if err = deleteLog(fileName); err != nil {
-				//log.Println("[LOGSERVER-Error] Unable to delete log.")
-				logger.XconfLogError("Unable to delete log.", true)
-
-				c.AbortWithStatusJSON(http.StatusNotFound, gin.H{
-					"message": "Delete error",
-					"reason":  err.Error(),
-				})
-				return
-			}
-
-			c.JSON(http.StatusOK, gin.H{
-				"message": "Delete success",
-				"reason":  "",
-			})
-		} else if u.Scheme == "http" {
-			//log.Println("[LOGSERVER-Info] Deleting log from " + u.Path + ".")
-			logger.XconfLogDebug("Deleting log from "+u.Path+".", true)
-			// Todo: retrieve logs from HTTP server.
-
-			c.AbortWithStatusJSON(http.StatusUnsupportedMediaType, gin.H{
-				"message": "URL parse error",
-				"reason":  "HTTP protocol not supported",
-			})
-		} else {
-			//log.Println("[LOGSERVER-Error] Unsupported source URL " + src + ".")
-			logger.XconfLogError("Unsupported source URL "+src+".", true)
-
-			c.AbortWithStatusJSON(http.StatusUnsupportedMediaType, gin.H{
-				"message": "URL parse error",
-				"reason":  "protocol not supported",
-			})
-		}
-	})
+	router.DELETE("/api/v1/logs/:name", DeleteLog)
 
 	// Handler for GET REST API, http://<ip_address>:<port>/api/v1/logs/info.
-	router.GET("/api/v1/logs/info", func(c *gin.Context) {
-		//log.Println("[LOGSERVER-Info] In info handler.")
-		logger.XconfLogDebug("In info handler.", true)
-
-		// Parse filter input.
-		timeID := c.Query("id")
-		fileName := c.Query("file_name")
-		fileSize := c.Query("size")
-		sizeLowerBound := c.Query("size.gt")
-		sizeUpperBound := c.Query("size.lt")
-		//location := c.Query("location")
-		owner := c.Query("owner")
-		createDate := c.Query("create_date")
-		createDateLowerBound := c.Query("create_date.gt")
-		createDateUpperBound := c.Query("create_date.lt")
-		//contact := c.Query("contact")
-		//description := c.Query("description")
-
-		var err error
-		var filter LogFilter
-
-		// Set filter parameters.
-		if timeID != "" {
-			var id gocql.UUID
-			id, err = gocql.ParseUUID(timeID)
-			if err != nil {
-				//log.Println("[LOGSERVER-Error] Unable to parse id", timeID, ".")
-				logger.XconfLogError("Unable to parse id "+timeID+".", true)
-				c.AbortWithStatusJSON(http.StatusInternalServerError, gin.H{
-					"message": "Info retrieval error",
-					"reason":  err.Error(),
-				})
-				return
-			}
-			filter.TimeID = id
-		}
-		filter.FileName = fileName
-		filter.SizeLower = -1
-		filter.SizeUpper = -1
-		if sizeLowerBound != "" {
-			filter.SizeLower, err = strconv.ParseInt(sizeLowerBound, 10, 64)
-			if err != nil {
-				//log.Println("[LOGSERVER-Error] Unable to parse size.gt", sizeLowerBound, ".")
-				logger.XconfLogError("Unable to parse size.gt "+sizeLowerBound+".", true)
-				c.AbortWithStatusJSON(http.StatusInternalServerError, gin.H{
-					"message": "Info retrieval error",
-					"reason":  err.Error(),
-				})
-				return
-			}
-		}
-		if sizeUpperBound != "" {
-			filter.SizeUpper, err = strconv.ParseInt(sizeUpperBound, 10, 64)
-			if err != nil {
-				//log.Println("[LOGSERVER-Error] Unable to parse size.lt", sizeUpperBound, ".")
-				logger.XconfLogError("Unable to parse size.lt "+sizeUpperBound+".", true)
-				c.AbortWithStatusJSON(http.StatusInternalServerError, gin.H{
-					"message": "Info retrieval error",
-					"reason":  err.Error(),
-				})
-				return
-			}
-		}
-		if fileSize != "" {
-			filter.SizeLower, err = strconv.ParseInt(fileSize, 10, 64)
-			filter.SizeUpper = filter.SizeLower
-			if err != nil {
-				//log.Println("[LOGSERVER-Error] Unable to parse size", fileSize, ".")
-				logger.XconfLogError("Unable to parse size "+fileSize+".", true)
-				c.AbortWithStatusJSON(http.StatusInternalServerError, gin.H{
-					"message": "Info retrieval error",
-					"reason":  err.Error(),
-				})
-				return
-			}
-		}
-		//filter.Location = location
-		filter.Owner = owner
-		if createDateLowerBound != "" {
-			t, _ := time.Parse(time.RFC3339, createDateLowerBound)
-			//log.Printf("[LOGSERVER-Info] createDateLowerBound = %s", t)
-			msg := fmt.Sprintf("createDateLowerBound = %s", t)
-			logger.XconfLogDebug(msg, true)
-			filter.CreateDateLower, err = time.Parse(time.RFC3339, createDateLowerBound)
-			if err != nil {
-				//log.Println("[LOGSERVER-Error] Unable to parse create_data.gt", createDateLowerBound, ".")
-				logger.XconfLogError("Unable to parse create_data.gt "+createDateLowerBound+".", true)
-				c.AbortWithStatusJSON(http.StatusInternalServerError, gin.H{
-					"message": "Info retrieval error",
-					"reason":  err.Error(),
-				})
-				return
-			}
-		}
-		if createDateUpperBound != "" {
-			t, _ := time.Parse(time.RFC3339, createDateUpperBound)
-			//log.Printf("[LOGSERVER-Info] createDateUpperBound = %s", t)
-			msg := fmt.Sprintf("createDateUpperBound = %s", t)
-			logger.XconfLogDebug(msg, true)
-			filter.CreateDateUpper, err = time.Parse(time.RFC3339, createDateUpperBound)
-			if err != nil {
-				//log.Println("[LOGSERVER-Error] Unable to parse create_data.lt", createDateUpperBound, ".")
-				logger.XconfLogError("Unable to parse create_data.lt "+createDateUpperBound+".", true)
-				c.AbortWithStatusJSON(http.StatusInternalServerError, gin.H{
-					"message": "Info retrieval error",
-					"reason":  err.Error(),
-				})
-				return
-			}
-		}
-		if createDate != "" {
-			t, _ := time.Parse(time.RFC3339, createDate)
-			//log.Printf("[LOGSERVER-Info] createDate = %s", t)
-			msg := fmt.Sprintf("createDate = %s", t)
-			logger.XconfLogDebug(msg, true)
-			filter.CreateDateLower, err = time.Parse(time.RFC3339, createDate)
-			filter.CreateDateUpper = filter.CreateDateLower
-			if err != nil {
-				log.Println("[LOGSERVER-Error] Unable to parse create_data", createDate, ".")
-				logger.XconfLogError("Unable to parse create_data "+createDate+".", true)
-				c.AbortWithStatusJSON(http.StatusInternalServerError, gin.H{
-					"message": "Info retrieval error",
-					"reason":  err.Error(),
-				})
-				return
-			}
-		}
-		//filter.Contact = contact
-		//filter.Description = description
-
-		var info []LogEntry
-		info, err = retrieveLogInfo(gSession, &filter)
-		if err != nil {
-			c.AbortWithStatusJSON(http.StatusInternalServerError, gin.H{
-				"message": "Info retrieval error",
-				"reason":  err.Error(),
-			})
-			return
-		}
-
-		var interfaceSlice []interface{} = make([]interface{}, len(info))
-		for i, entry := range info {
-			interfaceSlice[i] = entry
-		}
-
-		type InfoResponse struct {
-			Message string        `json:"message"`
-			Reason  string        `json:"reason"`
-			Reply   []interface{} `json:"info"`
-		}
-		var response InfoResponse
-		response.Message = "Info retrieval success"
-		response.Reason = ""
-		response.Reply = interfaceSlice
-
-		c.JSON(http.StatusOK, response)
-	})
+	router.GET("/api/v1/logs/info", GetInfo)
 
 	// Handler for REST API, http://<ip_address>:<port>/index. Web UI.
 	router.LoadHTMLGlob("templates/index.html")
@@ -669,7 +139,7 @@ func main() {
 		// Open a cassandra session.
 		session, err = openSession(cassandraHosts)
 		if err != nil {
-			log.Println("[LOGSERVER-Error]", err)
+			//log.Println("[LOGSERVER-Error]", err)
 			logger.XconfLogError(err.Error(), true)
 			return
 		}
